# app/api/digest/route.ts
import { prisma } from "@/prisma/db";
import { hashPassword } from "@/utils/auth";
import { comparePassword } from "@/utils/auth";
import { generateToken } from "@/utils/auth";
import { generateToken2 } from "@/utils/auth";

import { authorize } from "@/utils/auth";
import { NextResponse } from "next/server";
import { verifyToken } from "@/utils/auth";
import { Stats, stat } from "fs";

//returns all the threads comments polls published by the given user
export async function GET(request: any) {
    try {
        
        //As a user, I want to read a daily digest post generated by AI that summarizes the top discussions, recorded matches, and standings, allowing me to quickly catch up on recent events.
        //check if digest was created already for today
        const start = new Date();
        start.setHours(0,0,0,0);

        const end = new Date(start);
        end.setDate(end.getDate() + 1);
        
        const digestSearch = await prisma.digest.findFirst({
          where:{
            date:{
                gte: start,
                lt: end
            }
          }
        })
        if(digestSearch){
          return NextResponse.json({message: digestSearch.text}, {status: 200});
        }

        //if not create new one.
        const sinceYesterday = new Date();
        sinceYesterday.setDate(sinceYesterday.getDate() - 1);
        sinceYesterday.setHours(0,0,0,0);

        const topThreads = await prisma.comment.groupBy({
          by: ['threadId'],
          where: {
            date: {
              gte: sinceYesterday
            },
            isHidden:false
          },
          _count: {
            threadId: true,
          },
          orderBy: {
            _count: {
              threadId: 'desc'
            }
          },
          take: 3
        });

        var dataInput = ""
        var discussions = ""
        if(topThreads.length > 0){

          discussions = "The following are discussions. Comments may reply to parent comments:\n\n"
          for(var t of topThreads){
            var currThreadId = t.threadId
            const threadResult = await prisma.threads.findFirst({
              where:{
                tid:currThreadId
              }
            });
            if(threadResult){
              discussions += "Discussion title: '" + threadResult.title + "' Comments: \n"
              discussions += await pulldiscussion(currThreadId);
              discussions += "\n\n"
            }
            
          }
        }
        else{
          discussions = "<no discussions recorded, ignore discussions>"
        }

        //now query on matches from yesterday
        const yesterdaysMatches = await prisma.matches.findMany({
          where:{
            date:{
                gte: sinceYesterday
            },
            status:"FINISHED"
          },
          include:{
            homeTeam:true,
            awayTeam:true
          }
        })

        var matchesText = "Yesterdays match scores:\n"
        if(yesterdaysMatches.length > 0){
            var currMatch = ""
            for(var match of yesterdaysMatches){
              currMatch = match.homeTeam.name + " v.s " + match.awayTeam.name + ": " + match.homeScore + "-" + match.awayScore + "\n"
            }
            matchesText+=currMatch
        }
        else{
          matchesText += "No matches played yesterday!"
        }



        //now query on team standings
        const topTeams = await prisma.team.findMany({
          orderBy:{
            position:"asc"
          },
          take:3
          
        });
        var teamData = ""
        if(topTeams.length > 0){
          teamData += "Current leading teams:\n"

          for(var team of topTeams){
            const position = team.position
            const teamName = team.name
            teamData += teamName +" standing:" + position + "\n"
          }
        }
        else{
          teamData = "No teams currently playing!"

        }

        var discussionSummary = await digest(discussions)
        var result = discussionSummary.choices[0].message.content + "\n\n" + matchesText + "\n" + teamData
        
        //store it in db
        const digestStore = await prisma.digest.create({
          data:{
            text:result
          }
        });

        return NextResponse.json({message: result}, {status: 200});
    
    }
    catch(ex: any){
        console.log(ex.message)
        return NextResponse.json({error: "unexpected error"}, {status: 500});
    }
}

async function pulldiscussion(tid:number):Promise<string>{
  
  var discussion = ""
  const sinceYesterday = new Date();
  sinceYesterday.setDate(sinceYesterday.getDate() - 1);
  const comments = await prisma.comment.findMany({
            where:{
              date: {
                gte: sinceYesterday
              },
              threadId:tid,
              isHidden:false
            },
            include:{
              replies:true,
              parentComment:{
                select:{
                  text:true
                }
              }
            }
  });
  for(var comment of comments){
    var currstr = "Comment: "
    currstr += "'" + comment.text + "' \n"

    if(comment.parentComment){
      currstr += "Reply to:"
      currstr += "'"+comment.parentComment.text+"'\n"
      
    }
    currstr += "\n"
    discussion += currstr
  }

  return discussion;
}



async function digest(input:string){
    const today = new Date()
    const LLMinput = "Digest date:" +today.toDateString() +". The following is the top 3 discussions on my football website yesterday, please write an enthusiastic summary of these discussions as a daily digest for my website using at most 250 words, NOTHING MORE PLEASE: " + input
    const response = await fetch(
        "https://router.huggingface.co/v1/chat/completions",
        {
          method: "POST",
          headers: {
            "Authorization": "Bearer " + process.env.HF_API_KEY2,
            "Content-Type": "application/json"
          },
          body: JSON.stringify({
            "model": process.env.HF_SUM_MODEL,
            "messages": [
              {
                "role": "system",
                "content": "You are a helpful assistant that writes an enthusiastic daily sports digest."
              },
              {
                "role": "user",
                "content": LLMinput 
              }
            ],
            "max_tokens": 250,
            "temperature": 0.7
          })
        }
      );
    
    const data = await response.json();
    return data
    const text = await response.text();
    return text
}