# openapi.yaml
openapi: 3.0.0
info:
  title: API
  description: ""
  version: 1.0.0
servers:
- url: http://localhost:3000
- url: ""
paths:
...
  /api/admin/appeal/approve:
    put:
      tags:
      - Admin Tests
      summary: Approve Appeal
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/appeal_approve_body"
            example:
              uid: 2
      responses:
        "200":
          description: succesful
      security:
      - BearerAuth: []
...
  /api/admin/appeal:
    get:
      tags:
      - Admin Tests
      summary: Get all Appeals
      requestBody:
        content:
          application/json:
            schema:
              type: object
            example: {}
      responses:
        "200":
          description: succesful
      security:
      - BearerAuth: []
  /api/admin/report:
    get:
      tags:
      - Admin Tests
      summary: Get All Reports
      requestBody:
        content:
          application/json:
            schema:
              type: object
            example: {}
      responses:
        "200":
          description: succesful
      security:
      - BearerAuth: []
...
  /api/admin/report/approve:
    put:
      tags:
      - Admin Tests
      summary: Approve Report
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/report_approve_body"
            example:
              rid: 2
      responses:
        "200":
          description: succesful
      security:
      - BearerAuth: []
  /api/admin/report/dismiss:
    put:
      tags:
      - Admin Tests
      summary: Dismiss Report
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/report_dismiss_body"
            example:
              rid: 3
      responses:
        "200":
          description: succesful
      security:
      - BearerAuth: []
...
  /api/digest:
    get:
      tags:
      - daily digest
      summary: Get digest
      parameters:
      - name: Authorization
        in: header
        required: false
        schema:
          type: string
        example: "Bearer {{accessToken}}"
      responses:
        "200":
          description: succesful
...
components:
  schemas:
...
    appeal_approve_body:
      type: object
      properties:
        uid:
          type: integer
...
    report_approve_body:
      type: object
      properties:
        rid:
          type: integer
    report_dismiss_body:
      type: object
      properties:
        rid:
          type: integer
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer